# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Tesla Oracle API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/tesla/commands/{vehicleTokenId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a command to a Tesla vehicle using the provided vehicle token ID and command details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tesla"
                ],
                "summary": "Submit command to Tesla vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle token ID that must be set in the request path to identify the vehicle",
                        "name": "vehicleTokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Command details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.SubmitCommandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Command submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_tesla-oracle_internal_models.SubmitCommandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or vehicle does not belong to the authenticated user.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "Vehicle not found or failed to get vehicle by token ID.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error, including command submission failures.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
        "/v1/tesla/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get config params for frontend app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get tesla-related configuration parameters",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/tesla/telemetry/subscribe/{vehicleTokenId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Subscribes a vehicle for telemetry data using the provided vehicle token ID in the request path.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Subscribe vehicle for Tesla Telemetry Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle token ID that must be set in the request path to fetch vehicle details",
                        "name": "vehicleTokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully subscribed to vehicle telemetry.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or invalid developer license.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "Vehicle not found or failed to get vehicle by VIN.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "409": {
                        "description": "Vehicle is not ready for telemetry subscription.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error, including decryption or telemetry subscription failures.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
        "/v1/tesla/telemetry/unsubscribe/{vehicleTokenId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unsubscribes a vehicle from telemetry data using the provided vehicle token ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Unsubscribe vehicle from Tesla Telemetry Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OnboardingService Token ID",
                        "name": "vehicleTokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unsubscribed from telemetry data.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "OnboardingService not found or owner information is missing.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
        "/v1/tesla/telemetry/{vehicleTokenId}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initiates the data flow for a Tesla vehicle using the provided vehicle token ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tesla"
                ],
                "summary": "Start data flow for Tesla vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle token ID that must be set in the request path to fetch vehicle details",
                        "name": "vehicleTokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully started data flow for the vehicle.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or vehicle does not belong to the authenticated user.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "Vehicle not found or failed to get vehicle by token ID.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "409": {
                        "description": "Vehicle is not ready for telemetry subscription. Call GetStatus endpoint to determine next steps.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error, including decryption or fleet status retrieval failures.",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_DIMO-Network_tesla-oracle_internal_models.SubmitCommandResponse": {
            "type": "object",
            "properties": {
                "commandId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.SubmitCommandRequest": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}